+++ main.ts
@@ -8,11 +8,73 @@ import {
 import { render } from "pug";
 import { getAPI, isPluginEnabled } from "obsidian-dataview";
 import type { DataviewApi } from "obsidian-dataview";
+import { parseFragment } from "parse5";
+import type {
+  ChildNode,
+  TextNode,
+  Element,
+  Template,
+} from "parse5/dist/tree-adapters/default";
 import {
   DEFAULT_SETTINGS,
   PugTemplatePluginSettings,
   SettingTab,
 } from "./settings";
+
+function renderNodes(childNodes: ChildNode[], parent: Node) {
+  for (const node of childNodes) {
+    switch (node.nodeName) {
+      // Wrap text nodes in `<span>` tags.
+      case "#text":
+        createSpan({ text: (node as TextNode).value, parent });
+        break;
+
+      // Skipping these cases
+      case "#comment":
+      case "#documentType":
+      case "template":
+        break;
+
+      // Everything else is treated like an HTML element.
+      default:
+        const child = node as Element | Template;
+        const tagName = child.tagName as keyof HTMLElementTagNameMap;
+        const attr = Object.fromEntries(
+          child.attrs.map(({ name, value }) => [name, value]),
+        );
+
+        // We won't recurse when a child's children are only text nodes,
+        // because we can stringify them and set the text content instead of recursing.
+        if (child.childNodes.every((node) => node.nodeName === "#text")) {
+          createEl(tagName, {
+            text: child.childNodes
+              .map((node) => (node as TextNode).value)
+              .join("\n"),
+            attr,
+            parent,
+          });
+        } else {
+          const element = createEl(tagName, {
+            attr,
+            parent,
+          });
+          renderNodes(child.childNodes, element);
+        }
+        break;
+    }
+  }
+}
+
+function renderError(parent: Node, error: string) {
+  parent.empty();
+  const pre = createEl("pre", {
+    parent,
+  });
+  createEl("code", {
+    text: error,
+    parent: pre,
+  });
+}
 
 export default class PugTemplatePlugin extends Plugin {
   settings: PugTemplatePluginSettings;
@@ -61,10 +123,12 @@ export default class PugTemplatePlugin extends Plugin {
         }
       }
 
-      el.innerHTML = render(source, { fm, basedir, filename, dv });
+      const pugHtml = render(source, { fm, basedir, filename, dv });
+      const documentFragment = parseFragment(null, pugHtml, {});
+      renderNodes(documentFragment.childNodes, el);
     } catch (e) {
       console.error(e);
-      el.innerHTML = `<pre><code>${source}</code></pre><span>${e.toString()}</span>`;
+      renderError(el, e.toString());
     }
   }
 }